+++ src/main/test/game/datatype/shield/PlasmaShieldTest.java
@@ -21,8 +21,8 @@ public class PlasmaShieldTest {
     public void testShouldIncreaseShield(){
         PlayerData player = new PlayerData(0L, "P05", "Interceptor", false);
         
-        long initProtection = player.getShield().getProtection();
-        long initMaxProtection = player.getShield().getMaxProtectionValue();
+        double initProtection = player.getShield().getProtection();
+        double initMaxProtection = player.getShield().getMaxProtectionValue();
         Assert.assertEquals(initProtection, ShieldConfig.NORMAL_SHIELD_PROTECTION);
         Assert.assertEquals(initMaxProtection, ShieldConfig.NORMAL_SHIELD_PROTECTION);
